parallel test configuration in progress

+++ src/main/java/io/devnindo/service/testunit/module_example/dummy/DummyData.java
@@ -5,7 +5,7 @@ import io.devnindo.datatype.schema.DataBean;
 public class DummyData implements DataBean
 {
     String name;
-    Long sleepingTime;
+    Integer sleepingTime;
 
     String reply;
 
@@ -13,7 +13,7 @@ public class DummyData implements DataBean
         return name;
     }
 
-    public Long getSleepingTime() {
+    public Integer getSleepingTime() {
         return sleepingTime;
     }
